+++ src/helpers/index.ts
@@ -2,5 +2,6 @@
 export { default as getKeyCombination } from "./getKeyCombination";
 export { default as getReservedShortcutHint } from "./getReservedShortcutHint";
 export { default as isInputOrTextArea } from "./isInputOrTextArea";
+export { default as isMacPlatform } from "./isMacPlatform";
 export { default as keyParser } from "./keyParser";
 // @endindex